refactor: move services to internal package
