+++ internal/controller/auth.go
@@ -9,7 +9,7 @@ import (
 	"github.com/gin-gonic/gin"
 	"github.com/julianstephens/license-server/internal/controller/httputil"
 	"github.com/julianstephens/license-server/internal/model"
-	"github.com/julianstephens/license-server/service"
+	"github.com/julianstephens/license-server/internal/service"
 )
 
 type AuthRequest struct {
