+++ internal/controller/user.go
@@ -9,7 +9,7 @@ import (
 	jsoniter "github.com/json-iterator/go"
 	"github.com/julianstephens/license-server/internal/controller/httputil"
 	"github.com/julianstephens/license-server/internal/model"
-	"github.com/julianstephens/license-server/service"
+	"github.com/julianstephens/license-server/internal/service"
 )
 
 type ScopeRequest struct {
